requirements.txt
langchain
langchain-community
chromadb
sentence-transformers
groq
pypdf
python-dotenv
chromadb
groq
langchain
langchain-community
pypdf
python-dotenv
langchain
langchain-community
chromadb
groq
pypdf
python-dotenv
sentence-transformers
